Improve handling of *DUMPing a stream
The checks in *DUMP tried to spot a stream (ie. a file that can be opened with OS_Find but that OS_File 5 doesn't think is there), but in doing so never set the address or offset, so the hex values shown on the left of the screen would be whatever junk was in r2 & r3.
Reorder the tests so it starts at 0 and runs until EOF.
Make use of OSFile_ReadWithType rather than trying to deduce if it's an untyped file ourselves.
Tested with
  *DUMP random:
from SystemDevs-1_33.

Version 5.53. Tagged as 'Kernel-5_53'