[498] Fix memory corruption when shrinking appspace too far
If lazy task swapping is disabled, and appspace is an AMB node, it was
possible to map out the page at &7000 and trigger an abort by attempting
to shrink application space to a negative size, e.g. by using
"*ChangeDyn -RamFSSize" to create a RAM disc larger than the amount of
available RAM.

Fix it by sanity checking the DA/AMBNode size within AMBDAHandler before
passing on the request to the lower-level growpages/shrinkpages calls
(shrinkpages requires the input to be valid and will break if given a
negative size).

Version 6.63. Tagged as 'Kernel-6_63'